fix wsign warnings

+++ src/bulkcouplingrhs.hh
@@ -56,7 +56,7 @@ void addCouplingBulkRHS(VelocityDiscreteFunctionType& rhs,double gamma,const Cur
 
       // extract the associated intersection
       auto intersectionIt(bulkGridPart.ibegin(bulkEntity));
-      while(intersectionIt->indexInInside()!=faceLocalIdx)
+      while(static_cast<std::size_t>(intersectionIt->indexInInside())!=faceLocalIdx)
         ++intersectionIt;
       const auto intersection(*intersectionIt);
